requirements/dev.txt
Django==4.2.*
django-extensions==3.2.3
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
pygraphviz==1.13
django-auditlog==3.0.0
pillow==10.4.0
